fix: [2.6] cast float array contains literals (#51130)
pr: #51118
issue: #51069

This is the 2.6 backport for PR #51118 / issue #51069.

Summary:
- Fix ARRAY<FLOAT> literal narrowing for array contains paths in 2.6.
- Keep the 2.6 diff scoped to `Array.h`, `JsonContainsExpr.cpp`,
`UnaryExpr.cpp`, and `ExprArrayTest.cpp`.
- Add regression coverage for the reported float literals.

Verification from branch author:
- `clang-format-15` on touched C++ files.
- `git diff --check origin/2.6...HEAD --
internal/core/src/common/Array.h
internal/core/src/exec/expression/ExprArrayTest.cpp
internal/core/src/exec/expression/JsonContainsExpr.cpp
internal/core/src/exec/expression/UnaryExpr.cpp`.

Verification before PR creation:
- `git diff --check origin/2.6...buqian/bob/2.6/fix-51069-array-float`.

+++ internal/core/src/common/Array.h
@@ -443,7 +443,7 @@ class Array {
             case DataType::FLOAT: {
                 for (int i = 0; i < length_; i++) {
                     auto val = get_data<float>(i);
-                    if (val != arr2.array(i).float_val()) {
+                    if (val != static_cast<float>(arr2.array(i).float_val())) {
                         return false;
                     }
                 }
@@ -705,7 +705,7 @@ class ArrayView {
             case DataType::FLOAT: {
                 for (int i = 0; i < length_; i++) {
                     auto val = get_data<float>(i);
-                    if (val != arr2.array(i).float_val()) {
+                    if (val != static_cast<float>(arr2.array(i).float_val())) {
                         return false;
                     }
                 }
